Addressed reviewer comments
* Changed ZK based provider to use old class name and gave new Kafka Admin provider new name
* Added deprecated annotation to old ZK provider and additional javadoc
  explaining the deprecation.
* Switched to using the configured admin client timeout from the main config class
* Fixed Copy Right date
