+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/config/KafkaTopicConfigProvider.java
@@ -7,26 +7,18 @@ package com.linkedin.kafka.cruisecontrol.config;
 import com.google.gson.Gson;
 import com.google.gson.stream.JsonReader;
 import com.linkedin.kafka.cruisecontrol.KafkaCruiseControlUtils;
+import com.linkedin.kafka.cruisecontrol.config.constants.ExecutorConfig;
+import kafka.server.ConfigType;
+import kafka.zk.AdminZkClient;
+import kafka.zk.KafkaZkClient;
+import scala.collection.JavaConversions;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
-import java.time.Duration;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.stream.Collectors;
-import org.apache.kafka.clients.admin.AdminClient;
-import org.apache.kafka.clients.admin.Config;
-import org.apache.kafka.clients.admin.ConfigEntry;
-import org.apache.kafka.common.config.ConfigResource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 
 /**
@@ -38,96 +30,53 @@ import org.slf4j.LoggerFactory;
  *   }
  * </pre>
  *
+ * This class uses the Zookeeper based admin client that will be removed in Kafka 3.0. Therefore this class has been
+ * deprecated and will be removed in a future Cruise Control release. A new
+ * {@link com.linkedin.kafka.cruisecontrol.config.TopicConfigProvider} implementation using the Kafka Admin Client has
+ * been created ({@link com.linkedin.kafka.cruisecontrol.config.KafkaAdminTopicConfigProvider}) and can be set using the
+ * {@code topic.config.provider.class} configuration setting.
+ *
  */
+@Deprecated
 public class KafkaTopicConfigProvider implements TopicConfigProvider {
-
-  private static final Logger LOG = LoggerFactory.getLogger(KafkaTopicConfigProvider.class);
-
   public static final String CLUSTER_CONFIGS_FILE = "cluster.configs.file";
+  public static final String ZK_KAFKA_TOPIC_CONFIG_PROVIDER_METRIC_GROUP = "KafkaTopicConfigProvider";
+  public static final String ZK_KAFKA_TOPIC_CONFIG_PROVIDER_METRIC_TYPE = "GetAllActiveTopicConfigs";
+  private String _connectString;
+  private boolean _zkSecurityEnabled;
   private static Properties _clusterConfigs;
-  private AdminClient _adminClient;
 
   @Override
-  //TODO: This is just returning a parsed file? Why not the cluster config from the actual cluster?
   public Properties clusterConfigs() {
     return _clusterConfigs;
   }
 
   @Override
   public Properties topicConfigs(String topic) {
-    Config topicConfig = null;
-    ConfigResource topicResource = new ConfigResource(ConfigResource.Type.TOPIC, topic);
+    KafkaZkClient kafkaZkClient = KafkaCruiseControlUtils.createKafkaZkClient(_connectString,
+                                                                              ZK_KAFKA_TOPIC_CONFIG_PROVIDER_METRIC_GROUP,
+                                                                              ZK_KAFKA_TOPIC_CONFIG_PROVIDER_METRIC_TYPE,
+                                                                              _zkSecurityEnabled);
     try {
-      LOG.debug("Requesting details for topic '{}'", topic);
-      topicConfig = _adminClient
-              .describeConfigs(Collections.singletonList(topicResource))
-              .all()
-              .get(KafkaCruiseControlUtils.ADMIN_CLIENT_REQUEST_TIMEOUT_MS, TimeUnit.MILLISECONDS)
-              .get(topicResource);
-    } catch (InterruptedException e) {
-      LOG.error("The request for the configuration of topic '{}' was interrupted", topic);
-      e.printStackTrace();
-    } catch (ExecutionException e) {
-      LOG.error("The request for the configuration of topic '{}' failed", topic);
-      e.printStackTrace();
-    } catch (TimeoutException e) {
-      LOG.error("The request for the configuration of topic '{}' timed out", topic);
-      e.printStackTrace();
-    }
-
-    if (topicConfig != null) {
-      return convertTopicConfigToProperties(topicConfig);
-    } else {
-      LOG.error("The configuration for topic '{}' could not be retrieved", topic);
-      return new Properties();
+      AdminZkClient adminZkClient = new AdminZkClient(kafkaZkClient);
+      return adminZkClient.fetchEntityConfig(ConfigType.Topic(), topic);
+    } finally {
+      KafkaCruiseControlUtils.closeKafkaZkClientWithTimeout(kafkaZkClient);
     }
   }
 
   @Override
   public Map<String, Properties> allTopicConfigs() {
-    Map<ConfigResource, Config> topicConfigs = null;
+    KafkaZkClient kafkaZkClient = KafkaCruiseControlUtils.createKafkaZkClient(_connectString,
+                                                                              ZK_KAFKA_TOPIC_CONFIG_PROVIDER_METRIC_GROUP,
+                                                                              ZK_KAFKA_TOPIC_CONFIG_PROVIDER_METRIC_TYPE,
+                                                                              _zkSecurityEnabled);
     try {
-      LOG.debug("Requesting configurations for all topics");
-      topicConfigs = _adminClient
-              .listTopics()
-              .names()
-              .thenApply(
-                      topicNameSet -> _adminClient.describeConfigs(
-                              topicNameSet.stream().map(name -> new ConfigResource(ConfigResource.Type.TOPIC, name)).collect(Collectors.toList())
-                      ).all()
-              )
-              .get(KafkaCruiseControlUtils.ADMIN_CLIENT_REQUEST_TIMEOUT_MS, TimeUnit.MILLISECONDS)
-              .get(KafkaCruiseControlUtils.ADMIN_CLIENT_REQUEST_TIMEOUT_MS, TimeUnit.MILLISECONDS);
-    } catch (InterruptedException e) {
-      LOG.error("The request for the configuration of all topics was interrupted");
-      e.printStackTrace();
-    } catch (ExecutionException e) {
-      LOG.error("The request for the configuration of all topics failed");
-      e.printStackTrace();
-    } catch (TimeoutException e) {
-      LOG.error("The request for the configuration of all topics timed out");
-      e.printStackTrace();
+      AdminZkClient adminZkClient = new AdminZkClient(kafkaZkClient);
+      return JavaConversions.mapAsJavaMap(adminZkClient.getAllTopicConfigs());
+    } finally {
+      KafkaCruiseControlUtils.closeKafkaZkClientWithTimeout(kafkaZkClient);
     }
-
-    Map<String, Properties> propsMap = new HashMap<>();
-    if (topicConfigs != null) {
-      LOG.debug("Converting {} Topic Configs into Properties", topicConfigs.size());
-      for (Map.Entry<ConfigResource, Config> entry : topicConfigs.entrySet()) {
-        propsMap.put(entry.getKey().name(), convertTopicConfigToProperties(entry.getValue()));
-      }
-      LOG.debug("Topic Config conversion complete");
-    } else {
-      LOG.error("Topic configurations for all topics on the cluster could not be retrieved");
-    }
-    return propsMap;
-  }
-
-  private static Properties convertTopicConfigToProperties(Config config) {
-    Properties props = new Properties();
-    for (ConfigEntry entry : config.entries()) {
-      props.put(entry.name(), entry.value());
-    }
-    return props;
   }
 
   private void loadClusterConfigs(String clusterConfigsFile) throws FileNotFoundException {
@@ -146,10 +95,8 @@ public class KafkaTopicConfigProvider implements TopicConfigProvider {
 
   @Override
   public void configure(Map<String, ?> configs) {
-
-    KafkaCruiseControlConfig ccConfig = new KafkaCruiseControlConfig(configs);
-    _adminClient = KafkaCruiseControlUtils.createAdminClient(KafkaCruiseControlUtils.parseAdminClientConfigs(ccConfig));
-
+    _connectString = (String) configs.get(ExecutorConfig.ZOOKEEPER_CONNECT_CONFIG);
+    _zkSecurityEnabled = (Boolean) configs.get(ExecutorConfig.ZOOKEEPER_SECURITY_ENABLED_CONFIG);
     String configFile = KafkaCruiseControlUtils.getRequiredConfig(configs, CLUSTER_CONFIGS_FILE);
     try {
       loadClusterConfigs(configFile);
@@ -160,6 +107,6 @@ public class KafkaTopicConfigProvider implements TopicConfigProvider {
 
   @Override
   public void close() {
-    _adminClient.close(Duration.ofMillis(KafkaCruiseControlUtils.ADMIN_CLIENT_CLOSE_TIMEOUT_MS));
+    // nothing to do.
   }
 }
